+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Minor changes and fixes
 
 * Fix ConverseJS build: translations were missing, and only english was available.
+* Replaced a sync file operation by an async one, to avoid blocking the server.
 
 ## 8.2.1