+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 * Fix ConverseJS build: translations were missing, and only english was available.
 * Replaced a sync file operation by an async one, to avoid blocking the server.
+* Fix ConverseJS bug: room information not correctly refreshed when modifications are made by other users
 
 ## 8.2.1