+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 * Fix ConverseJS build: translations were missing, and only english was available.
 * Replaced a sync file operation by an async one, to avoid blocking the server.
 * Fix ConverseJS bug: room information not correctly refreshed when modifications are made by other users.
-* Translation update: German.
+* Translation update: German, French.
 
 ## 8.2.1